combine networking stuff  to one module
