+++ src/stdio_utils.rs
